Move to ruby dir
