+++ ruby/example/test_words.rb
