+++ ruby/exe/vosk-transcriber
