+++ ruby/lib/vosk/ffi.rb
