+++ ruby/lib/vosk/progressbar.rb
