+++ ruby/lib/vosk/version.rb
