+++ ruby/sig/vosk.rbs
