+++ ruby/spec/vosk/model_spec.rb
