+++ ruby/spec/vosk_spec.rb
